requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
websockets>=11.0.0
aiohttp>=3.8.0
asyncio-mqtt>=0.13.0

# Exchange and crypto
requests>=2.28.0
cryptography>=40.0.0
ccxt>=4.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.11.0

# Cache
redis>=4.5.0

# Notifications
python-telegram-bot>=20.0.0
discord.py>=2.3.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Monitoring
colorlog>=6.7.0
prometheus-client>=0.16.0

# Web interface
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0
bandit>=1.7.0

# System monitoring
psutil>=5.9.0
memory-profiler>=0.61.0